requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.8.0
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pathlib
python-multipart==0.0.6
